Add tests for Project.addDocument()
