+++ src/project/Project.js
@@ -1,10 +1,9 @@
 import { Component, DefaultDOMElement, platform } from 'substance'
 import { EditorPackage as TextureEditorPackage } from 'substance-texture'
-import { EMPTY_ARTICLE_XML } from '../article/articleHelpers'
 import SheetEditor from '../sheet/SheetEditor'
-import { generateEmptySheetXML } from '../sheet/sheetHelpers'
 import ProjectBar from './ProjectBar'
 import ContextPane from './ContextPane'
+import { addNewDocument } from './ProjectManipulations'
 
 export default class Project extends Component {
 
@@ -134,31 +133,7 @@ export default class Project extends Component {
   }
 
   _addDocument(type) {
-    let archive = this._getDocumentArchive()
-    let entries = archive.getDocumentEntries()
-    let name
-    let xml
-    // FIXME: Don't rely on vfs to be present for blank documents
-    if (type === 'sheet') {
-      let existingNames = new Set()
-      entries.forEach(e => {
-        if (e.type === 'sheet') {
-          existingNames.add(e.name)
-        }
-      })
-      name = `Sheet${existingNames.size+1}`
-      xml = generateEmptySheetXML(100, 26)
-    } else if (type === 'article') {
-      let existingNames = new Set()
-      entries.forEach(e => {
-        if (e.type === 'article') {
-          existingNames.add(e.name)
-        }
-      })
-      name = `Article${existingNames.size+1}`
-      xml = EMPTY_ARTICLE_XML
-    }
-    let newDocumentId = archive.addDocument(type, name, xml)
+    let newDocumentId = addNewDocument(type)
     this._openDocument(newDocumentId)
   }
