+++ tests/project/Project.test.js
@@ -1,5 +1,6 @@
 import test from 'tape'
 import ProjectTab from '../../src/project/ProjectTab'
+import { addNewDocument } from '../../src/project/ProjectManipulations'
 import createRawArchive from '../util/createRawArchive'
 import loadRawArchive from '../util/loadRawArchive'
 import { getSandbox } from '../testHelpers'
@@ -27,6 +28,25 @@ test('Project: validate resource name input on-the-fly', t => {
   t.end()
 })
 
+test('Project: add a new Sheet', t => {
+  let { archive, manifest } = _setup(sample())
+  let newId = addNewDocument(archive, 'sheet')
+  let entry = manifest.getDocumentEntry(newId)
+  t.notNil(entry, 'there should be a new entry')
+  t.equal(entry.type, 'sheet', '.. of type sheet')
+  t.equal(entry.name, 'Sheet2', '.. with name "Sheet2"')
+  t.end()
+})
+
+test('Project: add a new Article', t => {
+  let { archive, manifest } = _setup(sample())
+  let newId = addNewDocument(archive, 'article')
+  let entry = manifest.getDocumentEntry(newId)
+  t.notNil(entry, 'there should be a new entry')
+  t.equal(entry.type, 'article', '.. of type sheet')
+  t.equal(entry.name, 'Article2', '.. with name "Article2"')
+  t.end()
+})
 function sample() {
   return [
     {
